Add WebPlayer hoc, websocket client singleton
Added WebPlayer high order component, web player context to pass to its children, websocket client singleton, separated sdk and ws logic from MusicRoom component.
